requirements.txt
pandas
numpy
pathlib
python-docx